+++ rswag-specs/lib/rswag/specs.rb
@@ -12,6 +12,7 @@ module Rswag
       c.add_setting :swagger_root
       c.add_setting :swagger_docs
       c.add_setting :swagger_dry_run
+      c.add_setting :swagger_format
       c.extend ExampleGroupHelpers, type: :request
       c.include ExampleHelpers, type: :request
     end